refactor: 테이블 리팩터링
