+++ src/app/(admin)/_type/DetailTableType/DetailTableHeader.ts
@@ -1,0 +1,14 @@
+export interface TableHeaderItem {
+  key: string;
+  label: string;
+  className?: string;
+}
+
+interface TableData {
+  data: any;
+}
+
+export interface TableConfig {
+  headers: TableHeaderItem[];
+  data: TableData;
+}
